+++ owner/BUILD.bazel
@@ -32,10 +32,6 @@ go_test(
     ],
     embed = [":owner"],
     flaky = True,
-<<<<<<< HEAD:owner/BUILD.bazel
-=======
-    shard_count = 9,
->>>>>>> afdd5c2ecd5 (owner: fix data race on ownerManager.campaignCancel (#56362)):pkg/owner/BUILD.bazel
     deps = [
         "//ddl",
         "//infoschema",
